S0C-3A-1A: add shadow verification scenarios for chronicle entries and search index
- Introduced `shadow_verify_chronicle_entries` scenario to validate the integrity of chronicle entries against events in the database.
- Introduced `shadow_verify_search_index` scenario to ensure the consistency of search index entries for blocks, books, and tags.
- Updated `load_builtin_scenarios` to include the new scenarios.
- Modified failure-drills workflow to accommodate the new scenarios and ensure proper job distribution.
- Ensured backward compatibility with existing CLI commands while integrating new functionality.
